requirements.txt
# ===========================================
# MLOps Breast Cancer Project - Dependencies
# ===========================================

# Core ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Experiment Tracking
mlflow>=2.9.0

# Pipeline Orchestration
zenml>=0.55.0

# Hyperparameter Optimization
optuna>=3.5.0

# API & Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data Versioning
dvc>=3.30.0
dvc-gdrive>=3.0.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing & Code Quality
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
joblib>=1.3.0
requests>=2.31.0
